backend/requirements-dev.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
email-validator==2.3.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.11
python-dotenv==1.0.0
PyPDF2==3.0.1
httpx==0.25.2
redis==5.0.1
chromadb==0.4.18
sentence-transformers==2.2.2
python-docx==1.1.0
python-pptx==0.6.23
tiktoken==0.5.2
boto3==1.42.24
